Menambahkan Route disribusi pada Home
