+++ src/index.js
@@ -5,6 +5,7 @@ import Home from './Home';
 import reportWebVitals from './reportWebVitals';
 import HomeManager from './HomeManager';
 import Login from './Login/Login';
+import Distribusi from './Stylist/Distribusi';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
@@ -14,6 +15,7 @@ root.render(
       <Routes>
         <Route path="/" element={<Login />} />
         <Route path="/home/*" element={<Home />} />
+        <Route path="/distribusi/*" element={<Distribusi />} />
         <Route path="/manager" element={<HomeManager />} />
       </Routes>
     </BrowserRouter>
